Start preparing for Mingw build
While doing that, realize that cross build from macOS is not only not
advisable[*] but really not possible without the POSIX thread model.

[*] Doing that would mean compiling MK on MSYS2 referencing the libstdc++
on MSYS2 and then using the libstdc++ on macOS. Perhaps good, since GNU
does not bump the ABI every second, but still less safe for sure than doing
all the steps of the build using the same toolchain.